Use maven-rpm-plugin's defaults for RPM version and release.
Jenkins has a bug that breaks the formatting of maven.build.timestamp.
This means our RELEASE property would always introduce invalid RPM
release headers under a CI server.

We now have an empty RELEASE, for the situations when the user desires
a simpler RPM release.  Otherwise, the defaults of the Maven RPM
plugin are good for snapshot repositories.  And they will do exactly
what we'll need when running the release.