Define the RPM version and release properties early
Changes in 1.25 mean that during filtering of NCM components, the
RELEASE property would be empty.  In that case, we'd get bogus lines
like

"/software/packages" = pkg_repl("ncm-accounts", "7.0.0-", "noarch");

with a missing release.  Now we define rpm.version and rpm.release
earlier, and we'll replace all components and archetypes so that they
use rpm.release instead of RELEASE.

When specified in the command line, RELEASE will set rpm.release.